Add multilingual support
Add library for i18n in Angular.
Add list of words for French, Spanish and English.
Translate home view, add combo box for select the language and improve UI.
Create help view for display how to play instructions.
Move controllers to a separate file.
Define service for define the UI language.
Define service for get the words to use in the game.
Remove links in header.
[Time: 1h]